refactor(stella-tui): split the #4155 edit-attribution tests into their own file
model/tests.rs sat at 1497 of its 1500-line ceiling — three lines of
headroom, the tightest file in the tree, and ungrandfathered.

The seam is the one the file had already drawn: a banner reading
`#4155: a mutating row resolves the change its own call made`, fencing
six tests about which turn's edit claims which diff and what a row shows
when the answer is none. Those six move to `model/tests/edits.rs`,
beside the `producer_seq`, `retention` and `scope_review` modules that
directory already holds.

The four helpers under that banner — `edit_call`, `turn_boundary`,
`inline_ref`, `inline_refs` — did **not** move. `producer_seq` uses all
four, so they belong to the parent rather than to the block that
happened to sit above them; moving them was the first attempt and the
compiler said so.

No test was renamed, reordered or dropped: 44 before, 44 after, and
`cargo test -p stella-tui --lib model::` passes 91.

One doc comment changed wording. `check-prose`'s baseline is per (file,
pattern), so a grandfathered construction that moves to a new file stops
being grandfathered — CI caught `honest` in the moved block, and the rule
for that is to fix the prose rather than add a baseline entry. The
sentence now states the fact it was gesturing at: an edit reverted within
the same turn changed nothing on disk, so the row has nothing to show.

model/tests.rs 1497 -> 1342, with 158 lines of headroom. No baseline
entry: the file is not grandfathered and must not become so.

Third of the three files #5225 names — `drive.rs` and
`self_driving_cmd.rs` were the first two. `model.rs` itself is no longer
over: it reads 1439 on today's main, so the issue's table has moved on.

Closes #5225
